reorganize and shim ScanExecMeta overrides and fix interval file IO
